Conflicts:
	.gitignore
	SISCONDOC/nbproject/genfiles.properties
	SISCONDOC/nbproject/project.properties
	SISCONDOC/src/almacenamiento/accesodatos/BaseDatos.java
	SISCONDOC/src/almacenamiento/accesodatos/DAOAspirante.java
	SISCONDOC/src/almacenamiento/accesodatos/DAOConvocatoria.java
	SISCONDOC/src/almacenamiento/accesodatos/DAOUser.java
	SISCONDOC/src/almacenamiento/controlador/ControlAspirante.java
	SISCONDOC/src/almacenamiento/controlador/ConvocatoriaController.java
	SISCONDOC/src/almacenamiento/controlador/UserController.java
	SISCONDOC/src/presentacion/.goutputstream-I3B1OX
	SISCONDOC/src/presentacion/PanelDigitador.form
	SISCONDOC/src/presentacion/PanelDigitador.java
	SISCONDOC/src/presentacion/VistaAdmin.form
	SISCONDOC/src/presentacion/VistaAdmin.java
	SISCONDOC/src/presentacion/VistaConvocatoria.form
	SISCONDOC/src/presentacion/VistaConvocatoria.java
	SISCONDOC/src/presentacion/VistaCrearUsuario.form
	SISCONDOC/src/presentacion/VistaCrearUsuario.java
	SISCONDOC/src/presentacion/VistaEditarUsuario.form
	SISCONDOC/src/presentacion/VistaEditarUsuario.java
	SISCONDOC/src/presentacion/VistaLogin.form
	SISCONDOC/src/presentacion/VistaLogin.java
	SISCONDOC/src/proceso/Aspirante.java
	SISCONDOC/src/proceso/Convocatoria.java
	SISCONDOC/src/proceso/Usuario.java
	scriprBD.sql
	src/almacenamiento/accesodatos/BaseDatos.java
	src/almacenamiento/accesodatos/DAOAspirante.java
	src/almacenamiento/accesodatos/DAOConvocatoria.java
	src/almacenamiento/accesodatos/DAOUser.java
	src/almacenamiento/controlador/ControlAspirante.java
	src/almacenamiento/controlador/ConvocatoriaController.java
	src/almacenamiento/controlador/UserController.java
	src/presentacion/PanelDigitador.java
	src/presentacion/Thumbs.db
	src/presentacion/VistaAdmin.form
	src/presentacion/VistaAdmin.java
	src/presentacion/VistaConvocatoria.form
	src/presentacion/VistaConvocatoria.java
	src/presentacion/VistaCrearUsuario.form
	src/presentacion/VistaCrearUsuario.java
	src/presentacion/VistaEditarUsuario.form
	src/presentacion/VistaEditarUsuario.java
	src/presentacion/VistaLogin.form
	src/presentacion/VistaLogin.java
	src/proceso/Aspirante.java
	src/proceso/Convocatoria.java
	src/proceso/Usuario.java

+++ SISCONDOC/src/proceso/Aspirante.java
@@ -20,7 +20,7 @@ package proceso;
     private String name;
     private String lastname;
     private String city;
-    private int puntaje;
+    private double puntaje;
     private Convocatoria call;
     private String genero;
     private String jornada;
@@ -33,7 +33,7 @@ package proceso;
     
     
     public Aspirante () {} 
-    public Aspirante (String aspiranteDocument , String aspiranteName , String aspiranteLastname, String aspiranteCity ,int puntajeAspirante, Convocatoria call,String gen, String jor, String nacimiento, Formacion [] formacionesAspirantes, FormacionTic [] formacionesTicAspirantes,FormadorTIC [] formadoresTic,CEspecifico conocimientosAspirante, Idioma []  idiomasAspirante)
+    public Aspirante (String aspiranteDocument , String aspiranteName , String aspiranteLastname, String aspiranteCity ,double puntajeAspirante, Convocatoria call,String gen, String jor, String nacimiento, Formacion [] formacionesAspirantes, FormacionTic [] formacionesTicAspirantes,FormadorTIC [] formadoresTic,CEspecifico conocimientosAspirante, Idioma []  idiomasAspirante)
     {
         document = aspiranteDocument;
         name = aspiranteName;
@@ -110,14 +110,14 @@ package proceso;
     /**
      * @return the puntaje
      */
-    public int getPuntaje() {
+    public double getPuntaje() {
         return puntaje;
     }
 
     /**
      * @param puntaje puntaje set
      */
-    public void setPuntaje(int puntaje) {
+    public void setPuntaje(double puntaje) {
         this.puntaje= puntaje;
     }
